migrate handler objects to handler directories
